+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add OpenCensus Java OC-Agent Trace Exporter.
 - Add OpenCensus Java OC-Agent Metrics Exporter.
 - Add ability to filter metrics collected from Dropwizard registry.
+- Add optional service name to the Instana trace exporter.
 
 ## 0.18.0 - 2018-11-27
 - Set the